añadiendo id a las tareas

+++ src/main/java/com/univalle/todo/services/ITareaService.java
@@ -10,5 +10,5 @@ public interface ITareaService {
     TareaDTO crearTarea(TareaDTO tarea);
     EditTareaDTO modificarTarea(EditTareaDTO tarea);
     List<TareaListDTO> listarTareas(Integer id);
-    void eliminarTarea(Long id);
+    void eliminarTarea(Integer id);
 }
